Use MUnit in CompileRoutesSuite
